+++ app/breakout-rooms/page.tsx
@@ -1,0 +1,9 @@
+import BreakoutGroupsCard from "@/components/breakout/breakout-groups-card"
+
+export default function Page() {
+  return (
+    <div className="flex flex-col gap-4">
+      <BreakoutGroupsCard />
+    </div>
+  )
+}
